remove kaybee run from kb-importer start script; collect cron logs

+++ docker/kb-importer/start.sh
@@ -1,9 +1,12 @@
 #!/bin/sh
 
-#kaybee update
 mkdir -p /kb-importer/data
 cd /kb-importer/data
-mv /kb-importer/kb-importer.jar /kb-importer/kaybee /kb-importer/data
+if [ -f /kb-importer/kb-importer.jar ]
+then
+  mv /kb-importer/kb-importer.jar /kb-importer/kaybee /kb-importer/data
+fi
+
 ./kaybee update --force
 
 #Adding certs
@@ -12,34 +15,21 @@ for cert in $certs; do
    keytool -import -alias $cert -storepass changeit -keystore /usr/lib/jvm/java-1.8-openjdk/jre/lib/security/cacerts -file /kb-importer/certs/$cert -noprompt
 done
 
-#run kaybee import for kaybeeconf-initial.yaml
-if [ ! -f /kb-importer/data/init ]
-then
-  echo `date` " Running Initial Kaybee Import"
-  #wait for the backend to start
-  sleep 120
-  ./kaybee pull -c ../conf/kaybeeconf-initial.yaml
-  echo `date` " Running Kaybee Merge"
-  ./kaybee merge -c ../conf/kaybeeconf-initial.yaml
-  echo `date` " Kaybee Merge Done"
-  ./kaybee export -t steady -c ../conf/kaybeeconf-initial.yaml
-  chmod +x steady.sh
-  sh steady.sh
-  touch /kb-importer/data/init
-  echo `date` " Kaybee Import Done"
-fi
+#Wait for backend to start
+#sleep 120
 
 #create a cron job kaybeeconf.yaml
 crontab -l > tmpcron
 if ! cat tmpcron | grep "kb-importer.sh"
 then
-    if [ -z "$KB_IMPORTER_CRON" ]
+    if [ -z "$KB_IMPORTER_CRON" ]	
     then
-      echo "0 0 * * * /kb-importer/kb-importer.sh" >> tmpcron
+      echo "0 0 * * * /kb-importer/kb-importer.sh >> /kb-importer/cron.log 2>&1" >> tmpcron
     else
-      echo "$KB_IMPORTER_CRON" " /kb-importer/kb-importer.sh" >> tmpcron
+      echo "$KB_IMPORTER_CRON" " /kb-importer/kb-importer.sh  >> /kb-importer/cron.log 2>&1" >> tmpcron
     fi
 fi
 crontab tmpcron
+echo "cron job created."
 rm tmpcron
-crond -f
+cron -f
